warnings solved, unsigned/signed comparison, int to float conversion mostly

+++ src/SOIL.c
@@ -1550,7 +1550,7 @@ unsigned int SOIL_direct_load_DDS_from_memory(
 	unsigned int flag;
 	unsigned int cf_target, ogl_target_start, ogl_target_end;
 	unsigned int opengl_texture_type;
-	int i;
+	unsigned int i;
 	/*	1st off, does the filename even exist?	*/
 	if( NULL == buffer )
 	{
@@ -1684,7 +1684,7 @@ unsigned int SOIL_direct_load_DDS_from_memory(
 			/*	compressed DDS, MIPmap size calculation is block based	*/
 			shift_offset = 2;
 		}
-		for( i = 1; i <= mipmaps; ++ i )
+		for( i = 1; i <= (unsigned int)mipmaps; ++ i )
 		{
 			int w, h;
 			w = width >> (shift_offset + i);
@@ -1716,7 +1716,7 @@ unsigned int SOIL_direct_load_DDS_from_memory(
 	/*	do this for each face of the cubemap!	*/
 	for( cf_target = ogl_target_start; cf_target <= ogl_target_end; ++cf_target )
 	{
-		if( buffer_index + DDS_full_size <= buffer_length )
+		if( buffer_index + DDS_full_size <= (unsigned int)buffer_length )
 		{
 			unsigned int byte_offset = DDS_main_size;
 			memcpy( (void*)DDS_data, (const void*)(&buffer[buffer_index]), DDS_full_size );
@@ -1744,7 +1744,7 @@ unsigned int SOIL_direct_load_DDS_from_memory(
 					DDS_main_size, DDS_data );
 			}
 			/*	upload the mipmaps, if we have them	*/
-			for( i = 1; i <= mipmaps; ++i )
+			for( i = 1; i <= (unsigned int)mipmaps; ++i )
 			{
 				int w, h, mip_size;
 				w = width >> i;
